moved svg generator into jagua-rs-base
