+++ jagua-rs-base/Cargo.toml
@@ -20,6 +20,7 @@ tribool = {workspace = true}
 slotmap = {workspace = true}
 float-cmp = {workspace = true}
 itertools = {workspace = true}
+svg = {workspace = true}
 geo-offset = { version = "0.4.0", optional = true }
 geo-types = { version = "0.7.16", optional = true }
