+++ jagua-rs/src/lib.rs
@@ -1,18 +1,21 @@
-// packaging all modules into a single large crate
+// Packaging all modules into a single large crate with features
 
 #[doc(inline)]
 pub use jagua_rs_base::*;
 
-/// Strip Packing Problem (SPP) module for `jagua-rs`.
-#[cfg(feature = "spp")]
-pub mod spp {
-    #[doc(inline)]
-    pub use jagua_rs_spp::*;
-}
+/// Variants of 2D irregular Cutting and Packing Problems
+pub mod problem_modules {
+    /// Strip Packing Problem (SPP) module for `jagua-rs`.
+    #[cfg(feature = "spp")]
+    pub mod spp {
+        #[doc(inline)]
+        pub use jagua_rs_spp::*;
+    }
 
-/// Bin Packing Problem (BPP) module for `jagua-rs`.
-#[cfg(feature = "bpp")]
-pub mod bpp {
-    #[doc(inline)]
-    pub use jagua_rs_bpp::*;
+    /// Bin Packing Problem (BPP) module for `jagua-rs`.
+    #[cfg(feature = "bpp")]
+    pub mod bpp {
+        #[doc(inline)]
+        pub use jagua_rs_bpp::*;
+    }
 }
