+++ lbf/src/main.rs
@@ -13,8 +13,8 @@ use lbf::config::LBFConfig;
 use lbf::io::cli::Cli;
 use lbf::io::json_output::JsonOutput;
 use lbf::io::layout_to_svg::s_layout_to_svg;
-use lbf::opt::lbf_opt_bpp::LBFOptimizerBP;
-use lbf::opt::lbf_opt_spp::LBFOptimizerSP;
+use lbf::opt::lbf_bpp::LBFOptimizerBP;
+use lbf::opt::lbf_spp::LBFOptimizerSP;
 use lbf::{EPOCH, LBFInstance, LBFSolution, io};
 use log::{error, info, warn};
 use mimalloc::MiMalloc;
