+++ lbf/src/opt/mod.rs
@@ -1,4 +1,4 @@
-pub mod lbf_opt_bpp;
-pub mod lbf_opt_spp;
+pub mod lbf_bpp;
+pub mod lbf_spp;
 pub mod loss;
 pub mod search;
